% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.ordicluster.R
\name{my.ordicluster}
\alias{my.ordicluster}
\title{Return matrix go.}
\usage{
my.ordicluster(ord, clstr.rslt, prune = 0, display = "sites",
  w = weights(ord, display), col = 1)
}
\arguments{
\item{ord}{An ordinaiton object.}

\item{clstr.rslt}{REsult from hclust based on same distance as ordination.}

\item{prune}{0; do not change.}

\item{display}{"sites," do not change.}

\item{w}{weights function; do not change.}

\item{col}{1; do not change.}
}
\value{
Matrix used by gg_ordicluster to compute line segments.
}
\description{
This is a copy of vegan::ordicluster except with the plotting
statement commented out.Not exported.
}
