% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veganCovEllipse.R
\name{veganCovEllipse}
\alias{veganCovEllipse}
\title{veganCovEllipse}
\usage{
veganCovEllipse(cov, center = c(0, 0), scale = 1, npoints = 100)
}
\arguments{
\item{cov}{Covariance extracted from ordination with vegan's ordiellipse function}

\item{center}{Coordinates for center of ellipse}

\item{scale}{Ellipse dimension in standard deviations}

\item{npoints}{Number of points to generate}
}
\value{
A dataframe for plotting ellipses around each group centroid.
}
\description{
Generates points for ellipse around treatment centroids
}
\author{
Jari Oksanen
}
