% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ordicluster.R
\name{gg_ordicluster}
\alias{gg_ordicluster}
\title{Ordicluter Plot with ggplot2}
\usage{
gg_ordicluster(ord, clstr.rslt, plot = TRUE)
}
\arguments{
\item{ord}{An ordination object}

\item{clstr.rslt}{Result from hclust based on same distance as ordination}

\item{plot}{A logical; defaults to TRUE}
}
\value{
A list the plot and data frames used for plotting.
}
\description{
Ordicluter Plot with ggplot2
}
\examples{
data("dune")
dune.bray <- vegdist(dune, method="bray")
ord <- cmdscale(dune.bray, k=nrow(dune)-1, eig=TRUE, add=TRUE)
clstr.rslt <- hclust(dune.bray, method="single")
gg_ordicluster(ord, clstr.rslt)

}
