% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_arrow.R
\name{scale_arrow}
\alias{scale_arrow}
\title{scale arrows to plot}
\usage{
scale_arrow(arrows, data, at = c(0, 0), fill = 0.75)
}
\arguments{
\item{arrows}{A dataframe with two columns giving coordinates from envfit result.}

\item{data}{A two column mata frame giving coordinates extracted from ordination with vegan scores function.}
}
\value{
Silently returns a scaled data fram for plottin arrows.
}
\description{
Scales envfit arrows to fit within 75% of ordination plot dimendions.
}
